fix error when upload image from wa

+++ app/src/main/java/com/miftah/sehaty/ui/screens/detail/DetailViewModel.kt
@@ -78,8 +78,8 @@ class DetailViewModel @Inject constructor(
         _detailState.value.foodAfterScan?.let {
             val sodiumInGrams = it.sodium / 1000.0
 
-            val total =
-                it.totalCarbs + it.dietaryFiber + it.protein + sodiumInGrams + it.sugars + it.totalFat
+            val total = it.totalCarbs + it.dietaryFiber + it.protein + sodiumInGrams + it.sugars + it.totalFat
+
             // Prevent division by zero
             if (total == 0.0) {
                 _detailState.value = _detailState.value.copy(
@@ -95,12 +95,12 @@ class DetailViewModel @Inject constructor(
             } else {
                 _detailState.value = _detailState.value.copy(
                     dataNutrientPercentage = NutrientPercentage(
-                        dietaryFiber = (it.dietaryFiber.toFloat() / total.toFloat()) * 100f,
-                        totalCarbs = (it.totalCarbs.toFloat() / total.toFloat()) * 100f,
-                        sodium = (sodiumInGrams.toFloat() / total.toFloat()) * 100f,
-                        sugars = (it.sugars.toFloat() / total.toFloat()) * 100f,
-                        protein = (it.protein.toFloat() / total.toFloat()) * 100f,
-                        totalFat = (it.totalFat.toFloat() / total.toFloat()) * 100f
+                        dietaryFiber = ((it.dietaryFiber / total) * 100).toFloat(),
+                        totalCarbs = ((it.totalCarbs / total) * 100).toFloat(),
+                        sodium = ((sodiumInGrams / total) * 100).toFloat(),
+                        sugars = ((it.sugars / total) * 100).toFloat(),
+                        protein = ((it.protein / total) * 100).toFloat(),
+                        totalFat = ((it.totalFat / total) * 100).toFloat()
                     )
                 )
             }
